Converted Gradio Files into binary to reduce temp files.
Gradio for some dumb reason forces temp files to be created even when in binary mode which the documents says wouldn't happen. This is even worse for the default "file" mode where it actually creates two identical temp files for no reason. This way reduces the temp files to half of what it was.